Move test utilities cloudgene.mapred.util to cloudgene.mapred.test (to distinguish from unit tests for cloudgene.mapred.util classes).
