+++ src/test/java/cloudgene/mapred/api/v2/users/RegisterUserTest.java
@@ -14,8 +14,8 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 
-import cloudgene.mapred.TestApplication;
-import cloudgene.mapred.util.TestMailServer;
+import cloudgene.mapred.test.TestApplication;
+import cloudgene.mapred.test.TestMailServer;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import io.restassured.RestAssured;
 import jakarta.inject.Inject;
