+++ src/test/java/cloudgene/mapred/jobs/WorkflowEngineTest.java
@@ -11,7 +11,7 @@ import java.util.Map;
 import cloudgene.mapred.jobs.workspace.IWorkspace;
 import org.junit.jupiter.api.Test;
 
-import cloudgene.mapred.TestApplication;
+import cloudgene.mapred.test.TestApplication;
 import cloudgene.mapred.core.User;
 import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.jobs.sdk.WorkflowContext;
