+++ src/test/java/cloudgene/mapred/jobs/WrongWorkspaceTest.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import cloudgene.mapred.jobs.workspace.IWorkspace;
 import org.junit.jupiter.api.Test;
 
-import cloudgene.mapred.TestApplication;
+import cloudgene.mapred.test.TestApplication;
 import cloudgene.mapred.core.User;
 import cloudgene.mapred.database.dao.JobDao;
 import cloudgene.mapred.database.dao.UserDao;
