--- src/test/java/cloudgene/mapred/util/LoginToken.java
@@ -1,25 +1,0 @@
-package cloudgene.mapred.util;
-
-public class LoginToken {
-
-	private String accessToken = "";
-
-	private String csrfToken = "";
-
-	public String getAccessToken() {
-		return accessToken;
-	}
-	
-	public void setAccessToken(String accessToken) {
-		this.accessToken = accessToken;
-	}
-
-	public String getCsrfToken() {
-		return csrfToken;
-	}
-
-	public void setCsrfToken(String csrfToken) {
-		this.csrfToken = csrfToken;
-	}
-
-}
